Fix exception when changing clothes in KKS studio (#12)
It seems to be due to the [Coordinate Load Option] not being well behaved, but it seems that the source code is not available over there, so please allow me to fix it here.
Also, the method of getting _ksox in Awake() worked well. However, I thought it was not a good idea to write code in Awake that depends on the behavior of other components (calls to AddComponent), so I decided to set it up in Reload in a limited way.